split actors.ts
