+++ src/actors/utils.ts
@@ -1,0 +1,6 @@
+export function actorNameToToolName(actorName: string): string {
+    return actorName
+        .replace(/\//g, '-slash-')
+        .replace(/\./g, '-dot-')
+        .slice(0, 64);
+}
