+++ src/index.ts
@@ -6,4 +6,5 @@ export { createExpressApp } from './server.js';
 export { ApifyMcpServer } from './mcp-server.js';
 export { getActorAutoLoadingTools, getActorDiscoveryTools } from './tools/index.js';
 export { addActorToTools, discoverActorsTool, getActorsDetailsTool, removeActorFromTools } from './tools/index.js';
-export { getActorsAsTools, searchActorsByKeywords } from './actors.js';
+export { getActorsAsTools } from './actors/tools.js';
+export { searchActorsByKeywords } from './actors/search.js';
